backend-python/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pymongo==4.13.2
dnspython==2.7.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyJWT==2.10.1
bcrypt==4.3.0
pypdf==5.9.0
scikit-learn==1.7.1
joblib==1.5.1
torch==2.8.0
transformers==4.55.0
datasets==4.0.0
accelerate==1.10.1
